Set up architecture for dialogue system
Created all of the game objects and scripts needed for dialogue system, not yet fully implemented. Logic for main loop and path branching not coded. Event based architecture established and all files created, all that is left to do is write the logic for the actual execution of progression and branching of dialogue paths and having them interface with the UI